pyproject.toml
[project]
name = "breast-umls-kg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=3.0.0",
    "pyarrow>=23.0.0",
    "pyyaml>=6.0.3",
    "tqdm>=4.67.3",
]
